Minimal changes to get more similar to geth - Rename symbols - Move non-geth code to blockchain_ext.go - Move fields around to be more similar to geth - Add text formatting to make it more similar to geth
